spyros_twitter/pyproject.toml
[tool.poetry]
name = "vincenzo-2"
version = "0.1.0"
description = ""
authors = ["Spyros Roum <spyrosr1@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1"
aiosqlite = "^0"
bitlyshortener = "^0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
